+++ src/test/java/org/jcodec/containers/mp4/AppleGpsTest.java
@@ -1,6 +1,5 @@
 package org.jcodec.containers.mp4;
 
-import org.jcodec.containers.mp4.boxes.Box;
 import org.jcodec.containers.mp4.boxes.MetaBox;
 import org.jcodec.containers.mp4.boxes.MovieBox;
 import org.jcodec.containers.mp4.boxes.NodeBox;
@@ -9,11 +8,8 @@ import org.junit.Test;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 public class AppleGpsTest {
 
@@ -21,32 +17,8 @@ public class AppleGpsTest {
     public void testAppleGps() throws IOException {
         MovieBox moov = MP4Util.parseMovie(new File("src/test/resources/applegps/gps1.mp4"));
         UdtaBox udta = NodeBox.findFirst(moov, UdtaBox.class, "udta");
-        Box gps = findGps(udta);
-        ByteBuffer data = getData(gps);
-        assertNotNull(data);
-        data.getInt(); //skip 4 bytes
-        byte[] coordsBytes = new byte[data.remaining()];
-        data.get(coordsBytes);
-        String latlng = new String(coordsBytes);
+        String latlng = udta.latlng();
         assertEquals("-35.2840+149.1215/", latlng);
-    }
-
-    static ByteBuffer getData(Box box) {
-        if (box instanceof Box.LeafBox) {
-            Box.LeafBox leaf = (Box.LeafBox) box;
-            return leaf.getData();
-        }
-        return null;
-    }
-
-    static Box findGps(UdtaBox udta) {
-        List<Box> boxes1 = udta.getBoxes();
-        for (Box box : boxes1) {
-            if (box.getFourcc().endsWith("xyz")) {
-                return box;
-            }
-        }
-        return null;
     }
 
     @Test
